core/requirements.txt
asgiref==3.9.1
celery==5.4.0
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Deprecated==1.2.18
Django==5.2.6
django-celery-results==2.5.1
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-filter==25.1
django-unfold==0.68.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
lxml==6.0.2
packaging==25.0
pdfminer.six==20250506
pikepdf==9.10.2
pillow==11.3.0
pycparser==2.23
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
redis==5.0.7
uritemplate==4.2.0
wrapt==1.17.3
